fix(types): resolve all 38 mypy strict-mode errors
- pyproject.toml: add types-PyYAML to dev deps; remove unused overrides
  (webrtcvad, zmq, colorama, soundfile); add PIL.* override
- ears.py: add type annotations throughout; fix implicit Optional; update
  __main__ block to use current start() API (event_queue not callback)
- scribe.py: annotate transcribe() — explicit Optional + return type
- speaker.py: suppress no-any-return for scipy resample (returns Any)
- mouth.py: remove unused type: ignore for kokoro_onnx import
- events.py: change VisemeEvent.start_ms/duration_ms int → float
  (Kokoro produces sub-ms precision; tests were already using approx floats)
- config_schema.py: add Any import; add type args to FIELD_META dict
- metrics.py: add Any import; add type args to _SnapshotValue alias
- config_runtime.py: add type args to _coerce_value meta param
- handshake.py: introduce _Sendable Protocol; fix transport: object → _Sendable;
  add type args to _try_parse_hello_ack and _consume_hello_ack
- orchestrator.py: change _run_scribe audio: object → Any for transcribe compat
- store.py: assert lastrowid is not None before ChunkRecord construction
- loader.py: remove redundant type: ignore (pypdf in overrides)
- embedder.py: remove redundant type: ignore comments (sentence_transformers)
- vision.py: remove redundant type: ignore comments (llama_cpp/PIL in overrides);
  annotate result: Any to suppress union-indexable error

+++ src/core/events.py
@@ -110,8 +110,8 @@ class VisemeEvent:
 
     utterance_id: str  # binds this viseme to its utterance
     phoneme: str  # IPA or ARPAbet phoneme string
-    start_ms: int  # offset from utterance start, milliseconds
-    duration_ms: int  # phoneme duration, milliseconds
+    start_ms: float  # offset from utterance start, milliseconds
+    duration_ms: float  # phoneme duration, milliseconds
 
 
 @dataclass(frozen=True)
